fix(new-relic): fix apollo-server-plugin type declaration resolution. update .gitignore with newrelic logfile
Fix type declaration resolution. Type declaration modules need to be in folders that resolve similarly to the packages themselves. There is special formatting for type declarations that map to packages that are in child folders.  For this case that meant that @newrelic/apollo-server-plugin had to be resolved through newrelic__apollo-server-plugin

KLL-1683